requirements.txt
absl-py==0.1.13
astor==0.6.2
backports.weakref==1.0.post1
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
django-cors-headers==2.2.0
djangorestframework==3.7.7
djangorestframework-simplejwt==3.2.3
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
numpy==1.14.2
pandas==0.22.0
pbr==4.0.0
protobuf==3.5.2.post1
PyJWT==1.6.1
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
selenium==3.11.0
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
